Sort PostAsync test into correct test file
