+++ ShopifySharp.Tests/Services/Graph/GraphServiceTests.cs
@@ -5,12 +5,10 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using FakeItEasy;
 using FluentAssertions;
-using Newtonsoft.Json.Linq;
 using ShopifySharp.Infrastructure;
 using ShopifySharp.Infrastructure.Serialization.Http;
 using ShopifySharp.Tests.TestClasses;
 using Xunit;
-using JsonSerializer = Newtonsoft.Json.JsonSerializer;
 
 namespace ShopifySharp.Tests.Services.Graph;
 
@@ -92,55 +90,4 @@ public class GraphServiceTests
         await act.Should().ThrowAsync<TestException>();
     }
 
-    [Fact(DisplayName = "Lists orders using the GraphService")]
-    public async Task ListsOrdersUsingGraphService()
-    {
-        var query = @"
-              query listOrdersWithTag($limit: Int!) {
-                orders(first: $limit) {
-                  pageInfo {
-                    hasNextPage
-                    hasPreviousPage
-                  }
-                  edges {
-                    cursor
-                    node {
-                      id
-                      name
-                      tags
-                    }
-                  }
-                }
-              }
-            ";
-        var variables = new Dictionary<string, object>
-        {
-            { "limit", 10 }
-        };
-        // Serialize the GraphQL query and the variables into a JToken. Must use a JToken for now, or else the service
-        // will assume we are using a GraphQL string and send with the wrong content type.
-        var serializerSettings = Serializer.CreateNewtonsoftSettings();
-        var serializer = JsonSerializer.Create(serializerSettings);
-        var requestBody =  JToken.FromObject(new
-        {
-            query = query,
-            variables = variables
-        }, serializer);
-        // Send the request. For now this must be sent as a JToken, or else the service will assume it is a GraphQL
-        // string and send it with the wrong content type.
-        var jToken = await _sut.PostAsync(requestBody);
-        var listResult = jToken["orders"]?.ToObject<GraphListOrdersResult>();
-
-        Assert.NotNull(listResult);
-        Assert.NotNull(listResult.PageInfo);
-        Assert.NotEmpty(listResult.Edges);
-        Assert.All(listResult.Edges, edge =>
-        {
-            Assert.NotEmpty(edge.Cursor);
-            Assert.NotNull(edge.Node);
-            Assert.NotNull(edge.Node.Id);
-            Assert.NotNull(edge.Node.Name);
-            Assert.NotNull(edge.Node.Tags);
-        });
-    }
 }
